:arrow_up: #30 storybook を7系の最新安定版に更新
