--- .storybook/preview.js
@@ -1,15 +1,0 @@
-import { initialize, mswDecorator } from 'msw-storybook-addon';
-
-initialize();
-
-export const decorators = [mswDecorator];
-
-export const parameters = {
-  actions: { argTypesRegex: '^on[A-Z].*' },
-  controls: {
-    matchers: {
-      color: /(background|color)$/i,
-      date: /Date$/,
-    },
-  },
-};
