+++ package.json
@@ -14,8 +14,8 @@
     "lint:next": "next lint",
     "lint": "run-s lint:*",
     "format": "run-s fix:*",
-    "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook",
+    "storybook": "storybook dev -p 6006",
+    "build-storybook": "storybook build",
     "chromatic": "npx chromatic --project-token=$CHROMATIC_PROJECT_TOKEN"
   },
   "dependencies": {
@@ -41,15 +41,14 @@
   },
   "devDependencies": {
     "@babel/core": "^7.20.12",
-    "@storybook/addon-a11y": "^6.5.16",
-    "@storybook/addon-actions": "^6.5.16",
-    "@storybook/addon-essentials": "^6.5.16",
-    "@storybook/addon-interactions": "^6.5.16",
-    "@storybook/addon-links": "^6.5.16",
-    "@storybook/builder-webpack5": "^6.5.16",
-    "@storybook/manager-webpack5": "^6.5.16",
-    "@storybook/react": "^6.5.16",
-    "@storybook/testing-library": "^0.0.13",
+    "@storybook/addon-a11y": "^7.0.2",
+    "@storybook/addon-essentials": "^7.0.2",
+    "@storybook/addon-interactions": "^7.0.2",
+    "@storybook/addon-links": "^7.0.2",
+    "@storybook/blocks": "^7.0.2",
+    "@storybook/nextjs": "^7.0.2",
+    "@storybook/react": "^7.0.2",
+    "@storybook/testing-library": "^0.1.0",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@types/eslint": "^8.21.0",
@@ -74,11 +73,11 @@
     "eslint-plugin-testing-library": "^5.10.2",
     "jest": "^29.4.2",
     "jest-environment-jsdom": "^29.4.2",
-    "msw": "^1.0.1",
-    "msw-storybook-addon": "^1.7.0",
+    "msw": "^1.2.1",
+    "msw-storybook-addon": "^1.8.0",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.8.3",
-    "storybook-addon-next": "^1.7.1",
+    "storybook": "^7.0.2",
     "whatwg-fetch": "^3.6.2"
   },
   "msw": {
